+++ app/routes/devices.$id.$name.experiments.$experiment_id.tsx
@@ -71,6 +71,7 @@ export async function clientLoader({ serverLoader }: Route.ClientLoaderArgs) {
   return { experiments: experiments.data };
 }
 
+clientLoader.hydrate = true as const;
 export function HydrateFallback() {
   return <DefaultHydrateFallback />;
 }
