+++ app/routes/devices.$id.$name.experiments.tsx
@@ -64,6 +64,8 @@ export async function clientLoader({ serverLoader }: Route.ClientLoaderArgs) {
   return { experiments: experiments.data };
 }
 
+clientLoader.hydrate = true as const;
+
 export function HydrateFallback() {
   return <DefaultHydrateFallback />;
 }
