+++ app/routes/devices.$id.$name.hardware.$hardware_name.history.tsx
@@ -104,6 +104,8 @@ export async function clientLoader({
   return { data: sensorHistory?.data, events: deviceEvents?.data, ENV };
 }
 
+clientLoader.hydrate = true as const;
+
 export function HydrateFallback() {
   return <DefaultHydrateFallback />;
 }
