+++ app/routes/devices.$id.$name.tsx
@@ -131,6 +131,8 @@ export async function clientLoader({ serverLoader }: Route.ClientLoaderArgs) {
   };
 }
 
+clientLoader.hydrate = true as const;
+
 export function HydrateFallback() {
   return <DefaultHydrateFallback />;
 }
